requirements.txt
psycopg2-binary
SQLAlchemy
pandas
numpy
openpyxl
requests
python-dotenv